Refactor: import 정리 및 publicCompose 이름 수정
