+++ app/src/main/java/com/codehanzoom/greenwalk/MainActivity.kt
@@ -3,14 +3,6 @@ package com.codehanzoom.greenwalk
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import com.codehanzoom.greenwalk.PublicCompose.LargeButton
 import com.codehanzoom.greenwalk.ui.theme.GreenWalkTheme
 
 class MainActivity : ComponentActivity() {
