+++ app/src/main/java/com/codehanzoom/greenwalk/publicCompose/LargeButton.kt
@@ -1,11 +1,6 @@
-package com.codehanzoom.greenwalk.PublicCompose
+package com.codehanzoom.greenwalk.publicCompose
 
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.safeContentPadding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
@@ -13,41 +8,37 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.modifier.modifierLocalMapOf
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.codehanzoom.greenwalk.ui.theme.GreenWalkTheme
 
 @Composable
-fun SmallButton(title: String) {
+fun LargeButton(title: String ) {
     Button(
         shape = RoundedCornerShape(10.dp),
         colors = ButtonDefaults.buttonColors(
             containerColor = Color(0xff8CB369)
         ),
         modifier = Modifier
-            .size(100.dp, 40.dp)
-            .padding(0.dp),
+            .size(320.dp, 50.dp),
         onClick = { /*TODO*/ }) {
         Text(
             text = title,
             fontFamily = FontFamily.Monospace,
             fontWeight = FontWeight.Bold,
-            fontSize = 14.sp,
+            fontSize = 26.sp,
             textAlign = TextAlign.Center,
         )
     }
 }
-
 @Preview(showBackground = true)
 @Composable
-fun SmallButtonPreview() {
+fun LargeButtonPreview() {
     GreenWalkTheme {
-        SmallButton("기부하기")
+        LargeButton("시작하기")
     }
 }
