+++ app/src/main/java/com/codehanzoom/greenwalk/publicCompose/MaxWidthButton.kt
@@ -1,39 +1,33 @@
-package com.codehanzoom.greenwalk.PublicCompose
+package com.codehanzoom.greenwalk.publicCompose
 
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CutCornerShape
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.CornerRadius
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.codehanzoom.greenwalk.ui.theme.GreenWalkTheme
 
+
 @Composable
-fun LargeButton(title: String ) {
+fun MaxWidthButton(title: String ) {
     Button(
         shape = RoundedCornerShape(10.dp),
         colors = ButtonDefaults.buttonColors(
             containerColor = Color(0xff8CB369)
         ),
         modifier = Modifier
-            .size(320.dp, 50.dp),
+            .fillMaxWidth()
+            .height(50.dp),
         onClick = { /*TODO*/ }) {
         Text(
             text = title,
@@ -46,8 +40,8 @@ fun LargeButton(title: String ) {
 }
 @Preview(showBackground = true)
 @Composable
-fun LargeButtonPreview() {
+fun MaxWidthButtonPreview() {
     GreenWalkTheme {
-        LargeButton("시작하기")
+        MaxWidthButton("시작하기")
     }
 }
