+++ app/src/main/java/com/codehanzoom/greenwalk/publicCompose/TopBar.kt
@@ -1,10 +1,8 @@
-package com.codehanzoom.greenwalk.PublicCompose
+package com.codehanzoom.greenwalk.publicCompose
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
